2_company-profile/requirements.txt
crewai
langchain-google-genai
ddgs
httpx
python-dotenv
pydantic>=2.0.0
beautifulsoup4
